style: rename hex.rs as bytes.rs
